+++ packages/react-ui/components/Select/tids.ts
@@ -1,0 +1,5 @@
+export const SelectDataTids = {
+  root: 'Select__root',
+  menu: 'Select__menu',
+  label: 'Select__label',
+} as const;